🌐 Hosting my own API to Vercel. Renamed "app" to "api".
