+++ server/api/routes/weatherRoutes.js
