+++ server/index.js
@@ -1,5 +1,5 @@
 require('dotenv').config(); // import once, for use of env variables thru-out application/api
-const app = require('./app'); // main app >>> All dev inside this folder
+const app = require('./api'); // main app >>> All dev inside this folder
 const connectDB = require('./app/db/config'); // require the config file 
 
 connectDB(); // connect to database
