+++ server/vercel.json
@@ -9,7 +9,7 @@
     "routes": [
         {
             "src": "/api/(.*)",
-            "dest": "/server/app/index.js"
+            "dest": "/server/index.js"
         }
     ]
 }
